refactor: Use Assembler to replace the original converter
