+++ src/main/java/run/halo/app/controller/content/api/PostController.java
@@ -42,6 +42,7 @@ import run.halo.app.service.OptionService;
 import run.halo.app.service.PostCommentService;
 import run.halo.app.service.PostService;
 import run.halo.app.service.assembler.PostRenderAssembler;
+import run.halo.app.service.assembler.comment.PostCommentAssembler;
 
 /**
  * Content post controller.
@@ -57,6 +58,8 @@ public class PostController {
 
     private final PostService postService;
 
+    private final PostCommentAssembler postCommentAssembler;
+
     private final PostCommentService postCommentService;
 
     private final OptionService optionService;
@@ -66,10 +69,12 @@ public class PostController {
     private final PostAuthentication postAuthentication;
 
     public PostController(PostService postService,
+        PostCommentAssembler postCommentAssembler,
         PostCommentService postCommentService,
         OptionService optionService, PostRenderAssembler postRenderAssembler,
         PostAuthentication postAuthentication) {
         this.postService = postService;
+        this.postCommentAssembler = postCommentAssembler;
         this.postCommentService = postCommentService;
         this.optionService = optionService;
         this.postRenderAssembler = postRenderAssembler;
@@ -203,7 +208,7 @@ public class PostController {
             .listChildrenBy(postId, commentParentId, CommentStatus.PUBLISHED, sort);
         // Convert to base comment dto
 
-        return postCommentService.convertTo(postComments);
+        return postCommentAssembler.convertTo(postComments);
     }
 
     @GetMapping("{postId:\\d+}/comments/tree_view")
@@ -236,7 +241,7 @@ public class PostController {
         // Escape content
         postCommentParam.setContent(HtmlUtils
             .htmlEscape(postCommentParam.getContent(), StandardCharsets.UTF_8.displayName()));
-        return postCommentService.convertTo(postCommentService.createBy(postCommentParam));
+        return postCommentAssembler.convertTo(postCommentService.createBy(postCommentParam));
     }
 
     @PostMapping("{postId:\\d+}/likes")
