+++ src/main/java/run/halo/app/service/JournalCommentService.java
@@ -1,26 +1,15 @@
 package run.halo.app.service;
 
-import java.util.List;
-import org.springframework.data.domain.Page;
-import org.springframework.lang.NonNull;
-import org.springframework.lang.Nullable;
 import run.halo.app.model.entity.JournalComment;
-import run.halo.app.model.vo.JournalCommentWithJournalVO;
 import run.halo.app.service.base.BaseCommentService;
 
 /**
  * Journal comment service interface.
  *
  * @author johnniang
+ * @author guqing
  * @date 2019-04-25
  */
 public interface JournalCommentService extends BaseCommentService<JournalComment> {
 
-    @NonNull
-    List<JournalCommentWithJournalVO> convertToWithJournalVo(
-        @Nullable List<JournalComment> journalComments);
-
-    @NonNull
-    Page<JournalCommentWithJournalVO> convertToWithJournalVo(
-        @NonNull Page<JournalComment> journalCommentPage);
 }
