+++ src/main/java/run/halo/app/service/SheetCommentService.java
@@ -1,11 +1,6 @@
 package run.halo.app.service;
 
-import java.util.List;
-import org.springframework.data.domain.Page;
-import org.springframework.lang.NonNull;
-import org.springframework.lang.Nullable;
 import run.halo.app.model.entity.SheetComment;
-import run.halo.app.model.vo.SheetCommentWithSheetVO;
 import run.halo.app.service.base.BaseCommentService;
 
 /**
@@ -17,31 +12,4 @@ import run.halo.app.service.base.BaseCommentService;
  */
 public interface SheetCommentService extends BaseCommentService<SheetComment> {
 
-    /**
-     * Converts to with sheet vo
-     *
-     * @param comment comment
-     * @return a comment with sheet vo
-     */
-    @NonNull
-    SheetCommentWithSheetVO convertToWithSheetVo(@NonNull SheetComment comment);
-
-    /**
-     * Converts to with sheet vo
-     *
-     * @param sheetComments sheet comments
-     * @return a sheet comments with sheet vo
-     */
-    @NonNull
-    List<SheetCommentWithSheetVO> convertToWithSheetVo(@Nullable List<SheetComment> sheetComments);
-
-    /**
-     * Converts to with sheet vo
-     *
-     * @param sheetCommentPage sheet comments
-     * @return a page of sheet comments with sheet vo
-     */
-    @NonNull
-    Page<SheetCommentWithSheetVO> convertToWithSheetVo(
-        @NonNull Page<SheetComment> sheetCommentPage);
 }
