+++ src/main/java/run/halo/app/service/impl/JournalCommentServiceImpl.java
@@ -1,33 +1,22 @@
 package run.halo.app.service.impl;
 
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
-import org.springframework.util.Assert;
-import org.springframework.util.CollectionUtils;
 import run.halo.app.exception.NotFoundException;
-import run.halo.app.model.dto.JournalDTO;
-import run.halo.app.model.entity.Journal;
 import run.halo.app.model.entity.JournalComment;
-import run.halo.app.model.vo.JournalCommentWithJournalVO;
 import run.halo.app.repository.JournalCommentRepository;
 import run.halo.app.repository.JournalRepository;
 import run.halo.app.service.JournalCommentService;
 import run.halo.app.service.OptionService;
 import run.halo.app.service.UserService;
-import run.halo.app.utils.ServiceUtils;
+import run.halo.app.service.assembler.comment.JournalCommentAssembler;
 
 /**
  * Journal comment service implementation.
  *
  * @author johnniang
+ * @author guqing
  * @date 2019-04-25
  */
 @Service
@@ -39,8 +28,11 @@ public class JournalCommentServiceImpl extends BaseCommentServiceImpl<JournalCom
     public JournalCommentServiceImpl(JournalCommentRepository journalCommentRepository,
         OptionService optionService,
         UserService userService,
-        ApplicationEventPublisher eventPublisher, JournalRepository journalRepository) {
-        super(journalCommentRepository, optionService, userService, eventPublisher);
+        ApplicationEventPublisher eventPublisher,
+        JournalRepository journalRepository,
+        JournalCommentAssembler journalCommentAssembler) {
+        super(journalCommentRepository, optionService, userService, eventPublisher,
+            journalCommentAssembler);
         this.journalRepository = journalRepository;
     }
 
@@ -49,50 +41,5 @@ public class JournalCommentServiceImpl extends BaseCommentServiceImpl<JournalCom
         if (!journalRepository.existsById(journalId)) {
             throw new NotFoundException("查询不到该日志信息").setErrorData(journalId);
         }
-    }
-
-    @Override
-    @NonNull
-    public List<JournalCommentWithJournalVO> convertToWithJournalVo(
-        List<JournalComment> journalComments) {
-
-        if (CollectionUtils.isEmpty(journalComments)) {
-            return Collections.emptyList();
-        }
-
-        Set<Integer> journalIds =
-            ServiceUtils.fetchProperty(journalComments, JournalComment::getPostId);
-
-        // Get all journals
-        List<Journal> journals = journalRepository.findAllById(journalIds);
-
-        Map<Integer, Journal> journalMap = ServiceUtils.convertToMap(journals, Journal::getId);
-
-        return journalComments.stream()
-            .filter(journalComment -> journalMap.containsKey(journalComment.getPostId()))
-            .map(journalComment -> {
-                JournalCommentWithJournalVO journalCmtWithJournalVo =
-                    new JournalCommentWithJournalVO().convertFrom(journalComment);
-                journalCmtWithJournalVo.setJournal(
-                    new JournalDTO().convertFrom(journalMap.get(journalComment.getPostId())));
-                journalCmtWithJournalVo.setAvatar(buildAvatarUrl(journalComment.getGravatarMd5()));
-                return journalCmtWithJournalVo;
-            })
-            .collect(Collectors.toList());
-    }
-
-    @Override
-    @NonNull
-    public Page<JournalCommentWithJournalVO> convertToWithJournalVo(
-        @NonNull Page<JournalComment> journalCommentPage) {
-        Assert.notNull(journalCommentPage, "Journal comment page must not be null");
-
-        // Convert the list
-        List<JournalCommentWithJournalVO> journalCmtWithJournalVos =
-            convertToWithJournalVo(journalCommentPage.getContent());
-
-        // Build and return
-        return new PageImpl<>(journalCmtWithJournalVos, journalCommentPage.getPageable(),
-            journalCommentPage.getTotalElements());
     }
 }
