+++ src/main/java/run/halo/app/service/impl/SheetCommentServiceImpl.java
@@ -1,32 +1,18 @@
 package run.halo.app.service.impl;
 
-import static run.halo.app.model.support.HaloConst.URL_SEPARATOR;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
-import org.springframework.util.Assert;
-import org.springframework.util.CollectionUtils;
 import run.halo.app.exception.BadRequestException;
 import run.halo.app.exception.NotFoundException;
-import run.halo.app.model.dto.post.BasePostMinimalDTO;
 import run.halo.app.model.entity.Sheet;
 import run.halo.app.model.entity.SheetComment;
-import run.halo.app.model.enums.SheetPermalinkType;
-import run.halo.app.model.vo.SheetCommentWithSheetVO;
 import run.halo.app.repository.SheetCommentRepository;
 import run.halo.app.repository.SheetRepository;
 import run.halo.app.service.OptionService;
 import run.halo.app.service.SheetCommentService;
 import run.halo.app.service.UserService;
-import run.halo.app.utils.ServiceUtils;
+import run.halo.app.service.assembler.comment.SheetCommentAssembler;
 
 /**
  * Sheet comment service implementation.
@@ -45,8 +31,10 @@ public class SheetCommentServiceImpl extends BaseCommentServiceImpl<SheetComment
         OptionService optionService,
         UserService userService,
         ApplicationEventPublisher eventPublisher,
-        SheetRepository sheetRepository) {
-        super(sheetCommentRepository, optionService, userService, eventPublisher);
+        SheetRepository sheetRepository,
+        SheetCommentAssembler sheetCommentAssembler) {
+        super(sheetCommentRepository, optionService, userService, eventPublisher,
+            sheetCommentAssembler);
         this.sheetRepository = sheetRepository;
     }
 
@@ -59,87 +47,4 @@ public class SheetCommentServiceImpl extends BaseCommentServiceImpl<SheetComment
             throw new BadRequestException("该页面已被禁止评论").setErrorData(sheetId);
         }
     }
-
-    @Override
-    @NonNull
-    public SheetCommentWithSheetVO convertToWithSheetVo(@NonNull SheetComment comment) {
-        Assert.notNull(comment, "SheetComment must not be null");
-        SheetCommentWithSheetVO sheetCommentWithSheetVo =
-            new SheetCommentWithSheetVO().convertFrom(comment);
-
-        BasePostMinimalDTO basePostMinimalDto =
-            new BasePostMinimalDTO().convertFrom(sheetRepository.getOne(comment.getPostId()));
-
-        sheetCommentWithSheetVo.setSheet(buildSheetFullPath(basePostMinimalDto));
-
-        sheetCommentWithSheetVo.setAvatar(buildAvatarUrl(comment.getGravatarMd5()));
-
-        return sheetCommentWithSheetVo;
-    }
-
-    @Override
-    @NonNull
-    public List<SheetCommentWithSheetVO> convertToWithSheetVo(List<SheetComment> sheetComments) {
-        if (CollectionUtils.isEmpty(sheetComments)) {
-            return Collections.emptyList();
-        }
-
-        Set<Integer> sheetIds = ServiceUtils.fetchProperty(sheetComments, SheetComment::getPostId);
-
-        Map<Integer, Sheet> sheetMap =
-            ServiceUtils.convertToMap(sheetRepository.findAllById(sheetIds), Sheet::getId);
-
-        return sheetComments.stream()
-            .filter(comment -> sheetMap.containsKey(comment.getPostId()))
-            .map(comment -> {
-                SheetCommentWithSheetVO sheetCmtWithPostVo =
-                    new SheetCommentWithSheetVO().convertFrom(comment);
-
-                BasePostMinimalDTO postMinimalDto =
-                    new BasePostMinimalDTO().convertFrom(sheetMap.get(comment.getPostId()));
-
-                sheetCmtWithPostVo.setSheet(buildSheetFullPath(postMinimalDto));
-
-                sheetCmtWithPostVo.setAvatar(buildAvatarUrl(comment.getGravatarMd5()));
-
-                return sheetCmtWithPostVo;
-            })
-            .collect(Collectors.toList());
-    }
-
-    @Override
-    @NonNull
-    public Page<SheetCommentWithSheetVO> convertToWithSheetVo(
-        @NonNull Page<SheetComment> sheetCommentPage) {
-        Assert.notNull(sheetCommentPage, "Sheet comment page must not be null");
-
-        return new PageImpl<>(convertToWithSheetVo(sheetCommentPage.getContent()),
-            sheetCommentPage.getPageable(), sheetCommentPage.getTotalElements());
-    }
-
-    private BasePostMinimalDTO buildSheetFullPath(BasePostMinimalDTO basePostMinimalDto) {
-        StringBuilder fullPath = new StringBuilder();
-
-        SheetPermalinkType permalinkType = optionService.getSheetPermalinkType();
-
-        if (optionService.isEnabledAbsolutePath()) {
-            fullPath.append(optionService.getBlogBaseUrl());
-        }
-
-        if (permalinkType.equals(SheetPermalinkType.SECONDARY)) {
-            fullPath.append(URL_SEPARATOR)
-                .append(optionService.getSheetPrefix())
-                .append(URL_SEPARATOR)
-                .append(basePostMinimalDto.getSlug())
-                .append(optionService.getPathSuffix());
-        } else if (permalinkType.equals(SheetPermalinkType.ROOT)) {
-            fullPath.append(URL_SEPARATOR)
-                .append(basePostMinimalDto.getSlug())
-                .append(optionService.getPathSuffix());
-        }
-
-        basePostMinimalDto.setFullPath(fullPath.toString());
-        return basePostMinimalDto;
-    }
-
 }
